Added wasm target (#28)
* Added wasm target

* reverted change of chrono version

+++ src/tsid/display.rs
@@ -5,21 +5,21 @@ use std::fmt::{Display, Formatter};
 impl Display for TSID {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         let mut chars = String::with_capacity(13);
-        let number = self.number as usize;
+        let number = self.number;
 
-        chars.push(ALPHABET[(number >> 60) & 0b11111]);
-        chars.push(ALPHABET[(number >> 55) & 0b11111]);
-        chars.push(ALPHABET[(number >> 50) & 0b11111]);
-        chars.push(ALPHABET[(number >> 45) & 0b11111]);
-        chars.push(ALPHABET[(number >> 40) & 0b11111]);
-        chars.push(ALPHABET[(number >> 35) & 0b11111]);
-        chars.push(ALPHABET[(number >> 30) & 0b11111]);
-        chars.push(ALPHABET[(number >> 25) & 0b11111]);
-        chars.push(ALPHABET[(number >> 20) & 0b11111]);
-        chars.push(ALPHABET[(number >> 15) & 0b11111]);
-        chars.push(ALPHABET[(number >> 10) & 0b11111]);
-        chars.push(ALPHABET[(number >> 5) & 0b11111]);
-        chars.push(ALPHABET[number & 0b11111]);
+        chars.push(ALPHABET[((number >> 60) & 0b11111) as usize]);
+        chars.push(ALPHABET[((number >> 55) & 0b11111) as usize]);
+        chars.push(ALPHABET[((number >> 50) & 0b11111) as usize]);
+        chars.push(ALPHABET[((number >> 45) & 0b11111) as usize]);
+        chars.push(ALPHABET[((number >> 40) & 0b11111) as usize]);
+        chars.push(ALPHABET[((number >> 35) & 0b11111) as usize]);
+        chars.push(ALPHABET[((number >> 30) & 0b11111) as usize]);
+        chars.push(ALPHABET[((number >> 25) & 0b11111) as usize]);
+        chars.push(ALPHABET[((number >> 20) & 0b11111) as usize]);
+        chars.push(ALPHABET[((number >> 15) & 0b11111) as usize]);
+        chars.push(ALPHABET[((number >> 10) & 0b11111) as usize]);
+        chars.push(ALPHABET[((number >> 5) & 0b11111) as usize]);
+        chars.push(ALPHABET[(number & 0b11111) as usize]);
 
         return f.write_str(chars.as_str());
     }
